requirements.txt
# Core Framework Dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# AI and Semantic Kernel  
semantic-kernel>=1.37.0
openai>=1.6.1
tiktoken>=0.5.2

# Azure SDK Core
azure-core>=1.29.5
azure-identity>=1.15.0
azure-mgmt-core>=1.4.0

# Azure Cosmos DB
azure-cosmos>=4.5.1

# Azure Service Bus
azure-servicebus>=7.11.4


# Azure Storage (for logs and file handling)
azure-storage-blob>=12.19.0

# Azure Search (AI Search)
azure-search-documents>=11.4.0

# Azure Key Vault (for secrets management)
azure-keyvault-secrets>=4.7.0

# Data Processing
pandas>=2.1.4
numpy>=1.25.2
pypdf>=3.17.4

# HTTP and API
requests>=2.31.0
httpx>=0.25.2
aiohttp>=3.9.1

# Logging and Monitoring
structlog>=23.2.0
colorama>=0.4.6
termcolor>=2.4.0

# Development and Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
flake8>=6.1.0

# Email Processing (optional)
exchangelib>=5.0.3
imaplib2>=0.57

# Data Validation
marshmallow>=3.20.1
cerberus>=1.3.5

# Async Support
asyncio-mqtt>=0.16.1
aiofiles>=23.2.1

# JSON and Configuration
jsonschema>=4.20.0
pyyaml>=6.0.1
toml>=0.10.2

faker>=18.0.0